Avoiding failures when testing without cleaning in axis2-war.
git-svn-id: https://svn.apache.org/repos/asf/ode/branches/APACHE_ODE_1.1@642037 13f79535-47bb-0310-9956-ffa450edef68

+++ Rakefile
@@ -179,10 +179,12 @@ define "ode" do
     
     test.with projects("tools"), libs, AXIS2_TEST, AXIOM, JAVAX.servlet
     test.setup task(:prepare_webapp) do |task|
-      cp_r _("src/main/webapp"), _("target/test-classes")
-      cp Dir[_("src/main/webapp/WEB-INF/classes/*")], _("target/test-classes")
-      cp Dir[project("axis2").path_to("src/main/wsdl/*")], _("target/test-classes/webapp/WEB-INF")
-      cp project("bpel-schemas").path_to("src/main/xsd/pmapi.xsd"), _("target/test-classes/webapp/WEB-INF")
+      unless File.exist?(_("target/test-classes/webapp"))
+        cp_r _("src/main/webapp"), _("target/test-classes")
+        cp Dir[_("src/main/webapp/WEB-INF/classes/*")], _("target/test-classes")
+        cp Dir[project("axis2").path_to("src/main/wsdl/*")], _("target/test-classes/webapp/WEB-INF")
+        cp project("bpel-schemas").path_to("src/main/xsd/pmapi.xsd"), _("target/test-classes/webapp/WEB-INF")
+      end
     end
     test.setup unzip(_("target/test-classes/webapp/WEB-INF")=>project("dao-jpa-ojpa-derby").package(:zip))
   end
